[PATCH] EFS set_blocksize() error handling
There doesn't seem to be any active maintainer for EFS: It's either
abandoned, or never required specific patches too badly (also quite
possible in this case). 

It went into the tree in 2.3.2 and from there to 2.6.0-test6 I see
nothing that would look like maintainer's update.  OTOH, it's read-only
and shouldn't be too badly broken (or hard to keep alive).

Outside of trunk (i.e.  in 2.4.16--) we have a backport of global 2.5
change (sb_bread()) and check for set_blocksize() failures (from Alan). 
This is the 2.6.x equivalent.