[PATCH] /proc/sys/auxv
From: Roland McGrath <roland@redhat.com>

This was supposed to be part of the recent mm_struct.saved_auxv[] patch.

The /proc addition is half the reason for the patch, and the more important
one (letting you debug live processes, while NT_AUXV in core dumps lets you
debug dead ones).  The patch below was supposed to be part of the original.