[PATCH] ia64: nable SN2 in generic builds
Now that it works, we can enable sn2 in generic builds.  This should not
only allow generic kernels to work on sn2, but also fix the build
problems people have been seeing with the qla1280 driver.  I tested a
generic kernel built with this patch on zx1 and it worked.