[PATCH] [RFC]  fix compile erros in ini9100 driver
The attached patch fixes the compile errors from the DMA and scsi_cmnd
next usage. It has been tested on bugzilla here:
http://bugzilla.kernel.org/show_bug.cgi?id=213

I was not sure about the variable casting in the driver, but this is how
the qlogicisp driver did it. The driver also still needs to be converted
to the new error handling.