[ARM] Ensure that /proc/uptime returns sensible figures.
When we set xtime at boot from the RTC, we weren't setting the
monotonic time offset.  This had the effect of making the uptime
rather large.

We get around this problem by using the do_settimeofday() to set
the current time.  do_settimeofday() knows about this issue, and
will apply the appropriate correction to the monotonic time offset
for us.